Added CSS & Fixed weird idiot stuff in HTML
- CSS changes
  - Results look less like crap
  - Submit button is no longer weird
  - Navbar is lowercase and Comfortaa is used for cleaner look

- HTML changes
  - All supported airlines are an option
  - Fixed rambly incorrect 'About' section
  - New icon displayed in the browser tab
  - Converted minilogo to transparent PNG

I think that't about it, gonna come back to this soon to add cookies and
and databases and all that as soon as I learn how.

Ok, thanks to the US government (Bureau of Transportation Statistics)
for all the data I got from them.

Finally, hope you enjoy looking at how awful the JS was future me!